+++ src/main/java/lotto/domain/Lottos.java
@@ -1,5 +1,7 @@
 package lotto.domain;
 
+import lotto.domain.dto.LottoResult;
+
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Stream;
