+++ src/main/java/lotto/domain/dto/LottoResult.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.dto;
 
 public class LottoResult {
